requirements.txt
django
django-colorfield
django-oauth-toolkit